feat: NOT WORKING - use only city as filter
